gestion de droit

+++ src/main/java/edu/fbansept/demo_e3_24_25/controller/ReservationController.java
@@ -4,10 +4,14 @@ import com.fasterxml.jackson.annotation.JsonView;
 import edu.fbansept.demo_e3_24_25.dao.ReservationDao;
 import edu.fbansept.demo_e3_24_25.model.Reservation;
 import edu.fbansept.demo_e3_24_25.model.Utilisateur;
+import edu.fbansept.demo_e3_24_25.security.IsEmploye;
+import edu.fbansept.demo_e3_24_25.security.IsPartenaire;
+import edu.fbansept.demo_e3_24_25.security.MyUserDetails;
 import edu.fbansept.demo_e3_24_25.view.AffichageReservation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -23,6 +27,7 @@ public class ReservationController {
 
     @GetMapping("/reservation/{id}")
     @JsonView(AffichageReservation.class)
+    @IsEmploye
     public ResponseEntity<Reservation> get(@PathVariable int id) {
 
         Optional<Reservation> optionalReservation = reservationDao.findById(id);
@@ -37,6 +42,7 @@ public class ReservationController {
 
     @GetMapping("/reservations")
     @JsonView(AffichageReservation.class)
+    @IsEmploye
     public List<Reservation> getAll() {
 
         return reservationDao.findAll();
@@ -44,6 +50,7 @@ public class ReservationController {
 
     @GetMapping("/reservation-at-date/{date}")
     @JsonView(AffichageReservation.class)
+    @IsEmploye
     public List<Reservation> getAllByDate(@PathVariable String date) {
 
         LocalDateTime dateTime = LocalDateTime.parse(date);
@@ -55,6 +62,7 @@ public class ReservationController {
 
     @PostMapping("/reservation")
     @JsonView(AffichageReservation.class)
+    @IsEmploye
     public ResponseEntity<Reservation> create(@RequestBody Reservation reservation) {
 
         reservation.setId(null);
@@ -66,7 +74,11 @@ public class ReservationController {
 
     @PostMapping("/reservation-place-journee")
     @JsonView(AffichageReservation.class)
-    public ResponseEntity<Reservation> createReservationJournee(@RequestBody Reservation reservation) {
+    @IsPartenaire
+    public ResponseEntity<Reservation> createReservationJournee(
+            @RequestBody Reservation reservation,
+            @AuthenticationPrincipal MyUserDetails userDetails
+    ) {
 
         reservation.setId(null);
 
@@ -77,11 +89,7 @@ public class ReservationController {
         reservation.setDateDebut(dateDebut);
         reservation.setDateFin(dateFin);
 
-        //A remplacer par la personne connectée
-        Utilisateur fauxUtilisateur = new Utilisateur();
-        fauxUtilisateur.setId(1);
-
-        reservation.setUtilisateur(fauxUtilisateur);
+        reservation.setUtilisateur(userDetails.getUtilisateur());
 
         reservation.setDateCreation(LocalDateTime.now());
 
@@ -92,7 +100,9 @@ public class ReservationController {
 
 
     @PutMapping("/reservation/{id}")
+    @IsPartenaire
     public ResponseEntity<Reservation> update(
+            @AuthenticationPrincipal MyUserDetails userDetails,
             @RequestBody Reservation reservation,
             @PathVariable int id) {
 
@@ -104,18 +114,39 @@ public class ReservationController {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
+        Reservation reservationBdd = optionalReservation.get();
+
+        //si l'email de l'utilisateur connecté est différent de l'email de la personne
+        //qui a créé cette reservation, sauf si c'est un admin
+        if(!reservationBdd.getUtilisateur().getRole().equals("ROLE_ADMIN") &&
+                !reservationBdd.getUtilisateur().getEmail().equals(userDetails.getUsername())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         reservationDao.save(reservation);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @DeleteMapping("/reservation/{id}")
-    public ResponseEntity<Reservation> delete(@PathVariable int id) {
+    @IsPartenaire
+    public ResponseEntity<Reservation> delete(
+            @PathVariable int id,
+            @AuthenticationPrincipal MyUserDetails userDetails) {
 
         Optional<Reservation> optionalReservation = reservationDao.findById(id);
 
         if (optionalReservation.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        Reservation reservationBdd = optionalReservation.get();
+
+        //si l'email de l'utilisateur connecté est différent de l'email de la personne
+        //qui a créé cette reservation, sauf si c'est un admin
+        if(!reservationBdd.getUtilisateur().getRole().equals("ROLE_ADMIN") &&
+                !reservationBdd.getUtilisateur().getEmail().equals(userDetails.getUsername())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
         reservationDao.deleteById(id);
